requirements.txt
Flask==3.1.0
SQLAlchemy==2.1.1
Flask-Migrate==4.0.5
python-socketio==5.11.0
python-engineio==4.8.1
Werkzeug==3.1.0
python-dotenv==1.0.0
